pyproject.toml
[project]
name = "drgt"
version = "0.1.0"
description = "Drug response prediction with Graph Transformer"
requires-python = ">=3.9, <3.12"
dependencies = [
    "numpy<2",
    "pandas",
    "torch",
    "torch-geometric",
    "scikit-learn",
    "tqdm",
    "pubchempy",
    "seaborn",
    "pyyaml",
    "packaging",
    "rdkit-pypi",
    "ipykernel"
]
